Support sending flash data in chunks
Can be avoided by setting a large chunk size. This is to avoid
overflow on serial transport. A recommendation for serial transport is
to use a chunk size slightly smaller than the UART RX buffer size (on
the PSoC side). Cypress' bootloader_host.exe uses 25 as chunk size -
let's do the same here.